filter the data a bit
We were filtering recipes that didn't have a building, but then we need to also filter the outputs of those recipes.  This gets rid of things like the rifle.  Then we need to filter any recipes whose items are orphaned.  This gets rid of a few biomass related recipes that you can do in constructors, but would never be able to fully automate and so aren't needed in the planner.