large speedup to solver
Almost all of the improvement was from the fma.  It both keeps operations down with fewer reduce calls, and the early out 0 check saves lots of time for us on sparse matrixes.  A huge portion of the time is spent in that one tight inner loop.

The optimization to div did show some small gains as well.