dictionary rework
The constant column is now the last one instead of the first one.  Pivoting now leaves the location of all non-pivoted variables the same, so matrix order is different.  Many operations work in place on the matrix, although bigrats are still immutable.

This gives a miniscule speedup, but the main point was more for maintainability.